Buffham module import & robo24 scaffolding (#42)
Add Buffham support for importing other buffham files.

Due to several items on the roadmap not being completed (the associated
Github issue documents this), the parser has a few quirks on needing the
file's transitive dependencies spelled out (may work with only 1 layer
of dependencies declared, didn't test). Note that the order of the
dependencies matters (e.g. if it differs from another file with the same
dependencies) as the generated request IDs are local to each file being
parsed, but the resultant code for each dependency is shared across
modules. This is avoidable as other features become implemented.

Also add scaffolding code for the robo 24 workshop presentation.
Creating a second project allowed for generalizing a lot of things,
which was nice.