Add some symbolic macros (#58)
While #56 is still correct that this is a waste of time, I did hit a
breakthrough in porting some of the macros to the new symbolic macro
format when a Bazel maintainer brought up the `configurable = False`
option for `attr.label[_list]` to get actual `Label` / `list[Label]`s
instead of `select` objects.

Re-write some of the macros and reap no meaningful benefits. Some of the
macros (particularly Python ones) are impossible to port because an
upstream macro is both legacy (can't inherit from it) and breaks naming
conventions internally (raises an error). Some of them are possible but
my patience ran out.