Wizaidry (#111)
Create a library release of my Wizaidry project idea. The provided
library allows for the simple creation of proto-Agentic AI by supplying
it with Python functions it can invoke. Fully-featured handlers can
allow for text-to-text or voice-to-voice interactions.

On a private branch, this library was utilized to make an end-to-end
Arduino assistant. Purely through voice instruction, the AI wrote,
compiled, and uploaded an Arduino sketch to bring to the life the simple
circuit on my desk. The corresponding tools & prompts will be kept
private for now.

This also preps PyPi publishing for anyone seeking to use it. I'll try
and write a blog post on this soon.

Closes #110