Wizaidry arg fix (#113)
Add a small fix to #111 to not convert argument names to title case.
While that does match the JSON Schema behavior (at least as `pydantic`
defines it), it was confusing the models as they'd semi-randomly pick
between the original and the title-cased name. A few test runs of only
using the correct names yielded more consistent results.