Migrate to MCP tooling (#133)
Unknown to me at the time of making Wizaidry (which was, fundamentally,
a voice-to-voice demo),
[MCP](https://modelcontextprotocol.io/introduction) has become the
standard for tool invocations. I played around with a lot of MCP clients
and determined that they are all worthless, however VsCode has recently
adopted tool support with MCP servers, so I'm using that one. I also
wanted to run things locally - qwen3:8b runs surprisingly well on a CPU
- but the out-of-the-box performance from my Copilot education
subscription is too big to pass up; the hosted models perform much, much
better.

This PR open sources the tooling I made for the voice-to-voice demo.
With the release of coding agents, my file edit tooling is suddenly
unimpressive and outdated. Truth to be told, the file edit tooling
included with VsCode's is substantially better, so I will probably
abandon the "sandboxing" I made in any real tooling and delete
`FileEdit`.