Add Stardocs and rules release automation (#184)
Follow https://blog.aspect.build/stardocs-on-bcr to setup Stardocs on
rules. `rules_cc` and `pico_sdk` are lagging a bit, but most rules and
macros seem to be working.

It seems that the easiest way to publish these docs is to set up a full
release pipeline with:
- https://github.com/bazel-contrib/publish-to-bcr
-
https://github.com/bazel-contrib/.github/blob/master/.github/workflows/release_ruleset.yaml

so I guess we're doing that now.