Buffham enum support (#188)
Add simple `uint8_t` enum support to Buffham. While we're at it, start
the process of writing Buffham's schema with Buffham, allow in-line
comments for transactions & publishes, and allow publishes from other
namespaces.

While writing it in, I refactored most of the parser to have less code
duplication or nonsensical patterns going on.