Zero click auth with tsidp (#202)
[tsidp](https://github.com/tailscale/tsidp) from Tailscale is very neat.
It uses the authentication you already provided to log into your tailnet
to provide zero-click access to OIDC-compatible services.

While I like my Authentik setup for is configurability (a simple script
lets me share specific services with friends), the overhead of having to
either pay for an Identity Provider to make it "nice" or make my friends
set up yet-another-2FA mechanism + email verification makes it a barrier
for others to use it. Basically it has a pretty high initial setup cost
and a continual use cost (more 2FA codes 4ever).

tsidp shifts the burden elsewhere. At this stage, you still have the
burden of:
- your friends need to be logged into a tailscale account and configure
their device to use it
- you have to share each service + `tsidp` into their tailnet, which
requires a manually acceptance from them
  - not automated from my side either

But if you can meet that bar, everything else becomes zero-click
afterwards. I like it. It might only be cool for my tech-inclined
friends than say share-with-your-family, but hopefully Tailscale is
cooking up something more seamless to extend this.

As a plus (going all-in on always being on the tailnet), we can turn
funnel off on the service and keep things more secure. I didn't like
authing through an API key while it was funneled anyways.