Use setec for secrets management (#205)
I'm still a ways out from K8s life. But this weekend's project has a
simpler goal: clean up a lot of my docker compose secrets usage with a
better secrets manager.

My second take on `setec` was more successful this time. But it took a
lot of tooling to make it happen.

Closes #206 

TODOs:

- [x] Change more services to use setec
- [x] Rewrite `nlb/secrets/secrets.py` to be a setec client